Update issue 45 Having had a good look at this it never really worked well.
The runner managed the subscription list and not the agents class this means should we find
new agents during a load/reload it would never get subscribed to.  

I've now moved all the subscription management logic to the agents class and extended the 
connector plugins with an unsubscribe method.

The new logic is that on load/reload first the agents all get unsub'd, then new agents loaded
and finally re-subscribed.  We also had the give the PluginManager the ability to remove 
plugins so we can reload them.