Update issue 63:
Add a helper to create a authorization_hook:

module MCollective
    module Agent
        class Service<RPC::Agent
            authorized_by :foo


	    # ...
        end
    end
end


This will cause this plugin to be authorized by MCollective::Util::Foo.
The self.authorize(msg) method will get called with the entire raw msg
if this raises an exception the message will be denied else it will 
be processed as normal.

This makes it trivial for users to come up with their own wild schemes
for authorization.