Issue 83
RPC::Client had all kinds of place where it was either manipulating
filters directly and not really in safe ways either so nil's and other
nasties can get into the filters.

This patch changes all of the places except the main utility methods to
use that and the utility methods now compacts the arrays removing nils.