Issue 50 - Data Description Language DSL
Rework the whole DDL thing so that the meta data is stored
seperate from the agents.  The main motivation is to avoid
issues of having all the agents on the web ui box just to
get the DDL.  You can now just copy the ddl files onto
the node that wants access to it, not all the agents.

This patch removed the mc-rpc-help putting the functionality
into the main mc-rpc.

The client class now attempts to load the DDL and if finds it
it sets appropriate timeouts from the DDL.

The DDL class supports validating an action and arguments.
It gets used by the client to only make calls that would be
valid as described in the DDL.

Currently it validates:
 - all expected arguments are supplied
 - :string types the maxlength and regex
 - :list types that the supplied argument matches exactly

Everything else gets let through, we'll expand that in time.

Signed-off-by: R.I.Pienaar <rip@devco.net>