Issue 92 - Add display guidance to DDL
Add a RPC::Result object that mimics a hash by default but
that also knows which agent and action it belongs to.

Add a display handler to the DDL that operates on action level
that can indicate if we should always print, just print fails,
just print ok or flatten the results.

Update the old printrpc so it has a legacy mode and a new sexy
DDL enhanced mode where it will figure out on its own what
is is the right thing to do based on the new diplay properties
in the DDL

Signed-off-by: R.I.Pienaar <rip@devco.net>