Issue 5306 - SimpleRPC requests should support targeting a subset of nodes
Add the ability to limit the targets to a subset of discovered nodes.
You can limit either on a fixed number or based on a percentage like
10%.

It supports 2 methods of picking the subset:

  * first - picks the first result, a poor mans 'nearest' algorithm
  * random - picks any node randomly

You can set this behavior using the rpclimitmethod config option to
either first or random.

On the CLI you pass the following new switches:

  * -1 or --one - picks just 1 host based on the configured algorithm
  * --limit-nodes or --ln can take either a number or a percentage

In code you can set and query the limit using the limit_targets
method on the RPC client.

Signed-off-by: R.I.Pienaar <rip@devco.net>