5808 - Reloading Agent causes node to disappear
During the work to support multiple libdirs a check was added
to not reload agents that was already loaded giving the first
one found in many libdirs the priority.

This checked only worked the first time and not during future
invocations, to fix this we now delete plugins as well as unsub
from their broker topics.  This is more safe in general but also
means reloading now works as expected.