5832 - MC::Facts::Base doesnt properly force everything to strings
Fix the initial bug that was introduced during #5083

Also write a thorough test suite for the MC::Facts and MC::Facts::Base
classes.  During writing this a bug with MC::Facts::Base#has_fact? was
fixed.  A quick look through the code base suggests this is unused
code, still good to be fixed though.