5967 - Add the ability to create sub collectives
Introduce a concept to make subcollectives.  In the most basic
they are just different topic prefixes that a server can listen
on for messages.  Keeping track of the source of messages and
replying to the correct topic.

Plenty of changes were required on the client side to expose this
ability on the CLI and in code.

The inventory application is morphing into a more generic reporting
application as it can now report on collectives, draw dot graphs
of them etc.  We should rename it at some point