6392 - Create an exec helper
Create a simple way to vendor 3rd party libraries in a way
that allow distribution packages to remove our vendoring
and package the depndencies themselves.  This hopefully
strikes a balance between our need for a just-works solution
and the needs of the distributions.

Create MC::Shell that uses systemu to call external commands
and wait for them avoiding zombies.

Create MC::RPC::Agent#run thats a helper arond this for use
in agents.