6705 - Support implimenting actions using external scripts
Create the ability to delegate an action to an external script
in any language.

At the moment the only supported method is one using JSON as a
medium, more formats will be added specifically one to make writing
shell based systems easier.

This also includes a PHP helper class to make it easy to write
actions in PHP, this will be a basis and reference implimentation
from which other languages will be supported

We vendored the JSON gem to keep things simple for most users.