7711 - Validators raise UnknownRPCError instead of InvalidRPCData or MissingRPCData
Validators should raise InvalidRPCData or MissingRPCData but we caught
those exceptions and raised UnknownRPCError instead.

We now just ignore exceptions in the validation class - later on the
RPC::Agent#handlemsg will deal with unknown exceptions and do the right
thing.

Also tighten up the tests so they will validate this behavior in future
and add a test to test the case of missing data.