7755 - M::Applications#filter_extra_options can return an array and result in errors logged and MC_EXTRA_OPTS being ignored
In order to figure out where the config file live in the application framework
we parse MCOLLECTIVE_EXTRA_OPTS using shellwords for settings related to -c
or --config, if nothing is found related to config we should return an empty
string.

We didn't always return an empty string - in cases where there were no config
related settings we'd return an array and this would cause a STDERR error
to be printed.

The error is harmless but alarming, now we're returning a string as expected