7619 - Creating of destinations and messages should be left to the connectors
Previously the structure of the middleware was baked into the mcollective core
we had assumptions about how the middleware looked and this made it hard for
connector plugins to be effective.  Connector plugins had to abide by the
internal assumptions rather than have the freedom to impliment middleware however
they wanted.  We deliberately threw away message headers as received from the
middleware as they were considered too middleware specific.  This was a mistake
as effective use of the middleware relies on these.

We now have a MC::Message object that encapsulates all the properties of a mcollective
request or reply but in a way that does not assume a structure of the middleware.

The connector plugin now gets to make all decisions about the middleware structure.
All creation of topics, targets, subscriptions etc gets delegated to the connector

Additionally the Message object now takes care of validation, encoding, decoding
and base64 processing of messages as these functions only apply to message objects.

This simplified a lot of code especially in the main runner loop and improve out
ability to extend mcollective for the work related to #7226

The entire server and client logic has been reworked around this new object and
lots of dead code in the Util class was removed.

In the case where a Message object is actually a reply to an earlier Message the
earlier message is stored and the connector plugin will have access to the full
request that initiated a reply.  The connector will thus have access to all the
headers received and can make routing decision based on those.  The core has no
use for message headers, this is purely to support the connectors. This will make
doing #7246 and #6938 much eaier

This change also breaks backward compatability with early 1.1.x and earlier, this is
acceptable as the current production release is 1.2.x and we maintain compatability
with that.  The message protocol now has proper headers for the agent and collective
a message was sent to so the parsing of :msgtarget was retired.  So this header was
removed and old servers just wont understand these messages and we wont understand
old clients.