7988 - MCollective::Message should support point to point comms
This adds a new direct request method to the Message object
that will communicate with nodes directly.

With this feature enabled requests that goes out to a number
of nodes below a certain threshold (10 by default) will use
the new point to point mode while the rest will use broadcast.

This should be transparent to the end user.  But is disabled
by default

The idea with this is to enable communication with off-line
machines but also to optimize the communication where you need
to send a message to only 10 or so nodes and creating 10 messages
would be more efficient than sending a broadcast message to all of
1000s of nodes,

The other area where this is important is in the web space where
users can easily create arbitrary hosts lists by clicking on nodes.
In this scenario the grouping based on facts, classes etc does not
apply well.  Having a direct mode means you could supply arbitrary
hosts lists.

Ultimately this is the ground work for making discovery pluggable
and optional.  In the case of the web scenario above when the web
application provides a list of hostnames to communicate with the
mcollective infrastructure will not do a discovery but simply go
and communicate with those nodes.

This specific commit begins making the various internals aware of
this new mode, it adds 2 configuration options but leaves this mode
disabled by default.

The stomp connector will create queues per node to enable this,
doing it this way is pretty inefficient and wont work at scale
a new ActiveMQ connector will be created that use a much better
but ActiveMQ specific approach to solving the queues per server
problem.

Followup work needs to be done around request TTLs to avoid down
machines from doing lots of unintended work.