8461 - Improve support for compound_filters
During the work in #8181 a new filter type was introduced but
a few integration points were overlooked.  The result of these
were that for some kinds of requests no filter were applied.

This improves the Util#empty_filter method to return an empty
compound filter and also improves various aspects of the RPC::Client
class to be aware of the compound filter.