8703 - RPC::Agent#authorized_by loads auth plugins on each request
The authorized_by helper creates a method called authorization_hook and
loads up a class from disk that provides the implimentation as a plugin.

We used to only create 1 instance of a agent so this only happened once
but recently we started creating new instances of the agent on each
request which means this steps gets done many times.

This change avoids creating the hook if it's already defined thus avoiding
the undesired disk loads