8466 - When identities are supplied , do not do discovery
Allow the Message object to be of type :direct_request which would force
direct requests

Change identity filters to be and and not or since that makes more sense
for identities.  But also if we're sending messages with many identities
attached we'd only want to match if ours is anywhere in the list, an
and based selection would always fail.

If identities are provided on the CLI then we use those and avoid discovery.

The RPC::Client#discover function can now be passed a list of hosts in
either string, array or JSON formats.  It will extract from this data
discovery data and use that instead of the old discovery methods.  In
this use case it will force direct_addressing mode.

The rpc application can now take JSON on STDIN and will use the above
capabilities to fill in discovery data