10115 - providing hostlists for direct access mode doesnt work with large host lists
Previously when :auto was given to the call_agent it would do discover
and then force direct_requests off.

If discover was called before and an array was supplied this would
effectively overrule that setting and so cause a broadcast to be done
when it was not supposed to be done.

Now we don't force that since discover will do the right thing and
force direct requests if it feels thats whats needed