10327 - callerids and identitied should only allow a limited set of characters
To be certain we never attempt to read/write files we shouldn't for
example in naive registration plugins the framework now ensures that
callerids and identities can only be /\w\.\-/

There's a new method valid_callerid? on the security base class so
if there's some weird and wonderful need a security plugin could
override this check but generally the current setting should be
left alone as is.