7899 - Create an ActiveMQ specific plugin
Add an ActiveMQ connector, tests and docs.

This connector use the following queues and topics:

 - A topic per agent and collective - <collective>.<agent name>.agent
 - Replies goes into short lived queues - <collective>.replies.<unique string>
 - Each node subscribe to a queue with a JMS selector matching its
   identify - <collective>.nodes

The model of replies require the use of gcInactiveDestinations in
ActiveMQ to purge old unused short lived queues.  The hope is that we
will use actual temp queues in future but there's some issues
configuring these reliably in ActiveMQ

Each node subscribes to the nodes queue and use a JMS selector, this
requires a queue specific connection for broker to broker comms that
has conduitSubscriptions disabled.