10883 - Syntax errors in Applications not correctly parse and displayed to the user
Change the behavior of the PluginManager to raise errors as well as log
errors when a file fails to load.  This behavior is optional so we're
able to maintain old behavior in places where getting exceptions might
be problematic.

The Applications#run method now show a friendly error and backtrace
which would hopefully help users get to the bottom of the problem