11251 - mcollective `rake doc` fails on rdoc.gem > v2.3.0
Ruby 1.9 removed the html template for rdoc instead using one called
darkfish, the Rakefile specifically requests the use of the html
template

We didn't actually need to specify a specific template since rdoc does
the right thing by default on ruby 1.8.6 and newer.  So this commit
removes the explicit template name relying on the defaults.