11012 - Empty strings in RPC results doesnt display correctly with printrpc
The display helper would split a string on new lines and then print each
resulting line padded appropriately.

In the case of an empty line splitting on new lines would return an
empty array and so no output will be produced - not even a new line -
thus resulting in a badly formatted table.