12109 - Make the syslog facility configurable
Added the "logfacility" configuration key.   This allows users who are logging
to syslog the ability to set the desired facility.  Syslog defaults to LOG_USER
aka user when no facility is specified.

The value for "logfacility" is expected to be the shorthand common facility
names such as local5, daemon, mail, etc.  The "LOG_" prefix get's prepended
automatically when the syslog handle get's initialized.