12555 - Support for Windows Platform
Add initial support for Microsoft Windows.

The main issue is that we hard coded \tmp in a few places and used
Unix signal handlers that does not exist in windows

Windows has no Syslog and the syslog module for ruby is not in the
windows installers so we do a few tweaks to the syslog logger and
tests to just nil out in windows.  We should look into adding a
logger for the windows eventlog using win32-eventlog

In a few tests we used pwd and other unix utilities, we replace
all those to calls to a few ruby invocations that is more portable

Windows does not show colors the way unix does, rather than try to
create a windows compatible color output we just default to disabling
all color when run on windos

It's hard/impossible to figure out the correct terminal dimensions on
windows so the dynamic resizing progress bar behavior on Unix is not
reproducable - on windows we just default to 80 columns which is the
default for the terminal.

^C does not interrupt IO events on windows so we took some examples from
various large projects to just start a background thread that sleeps for
a second.  The effect of this is that the ruby VM has an oppertunity
once a second to process any ^C signals etc.

The help template defaults to the directory the cofig file is in and if
it cant find the help template there it falls back to the previous
defaults to maintain consistent behavior across versions.

A first stab at some support batch files to eventually facilitate
packaging is in ext/windows, this commit specifically does not take care
of any packaging requirements yet.  There's a quick README in that
directory to assist early adopters to test this code once released.
Actual packaging will happen in a different ticket/commit.