12597 - Create Plugin Packaging
MCollective plugin application allows us to package mcollective plugins as
debs or rpms using the mco command. Plugin types are defined by definition
classes that populates a list of files and metadata which are in turn passed
over to packager classes which will build the action package.

Currently agent plugins are defined in AgentDefinition and debs and rpms are created with the
OSpackagePackager, which uses FPM to output the packages.

Additionally the info action will display plugin related metadata and a list of packages
that will be created from a target directory.