12865 - the scroller should be shown earlier
We show a progress bar while results are gathered.  The progress bar is
only showed after the first result is received which might leave the
user wondering what is going on between creating the request and
receiving the first reply.  This might be a long time in the case of
package updates etc.

With this commit we show the progress bar with 0 progress before sending
the request