12927 - Application#application_failure should not handle SystemExit
Users can use exit() in applications as is expected in any normal ruby
script, the application_failure handler though gets exceptions and
prints backtraces.

In some cases the application class caught SystemExit and did not pass
those to the application_failure but it was only done in some cases.

With this commit the appliction_failure method now handles SystemExit
correctly by just raising it again after first disconnecting from the
middleware.  This avoids a few cases where users might get a error
printed to their display saying the application failed to run when in
fact all that happened is that it used exit()

Also adds test coverage to code that did not have it before