12852 - RPC::Client#discover should accept :nodes as :hosts
In the past we only accepted :hosts as a argument when supplying
discovery information.  The terminology choice was unfortunate as the
official terminology guide prefers the word 'node'

We now accept :nodes as well as :hosts and treat them as equivalent, the
documentation has been updated to prefer :nodes

Additionally the discover method will now raise an error if unknown
flags are passed in as it might lead to a user causing change against
all his nodes due to simple programming error, so being very strict here
is in the users best interests.