12950 - Message#base64_decode! and Message#base64_encode! is broken
In the past the message object used @body for the message payload, this
got changed later to be @payload.  The encode base64_encode! and
base64_decode! methods was never changed to reflect this change

Additionally the tests were not good enough to pick up that a nil was
being passed to the SSL library and not the correct payload.

This fixes both issues