12859 - Application should be able to disable common options
Certain types of application will not want the RPC, Filter or Common
options.  The plugin packager is a good example as it will never need to
connect to the middleware having those options will just confuse the
users.

A plugin can now elect to disable one of 3 groups of options.

Also add some test coverage to code that was not previous tested