13028 - Division by zero when no agents are discovered in verbose mode
This primarily fixes a bug introduced while working on #12865.  We
showed the progress bar earlier there to improve the user experience but
we didnt do sufficient checks to avoid showing it when nothing was
discovered.

Also fix some behavior on the rpc application where it would print the
call stats showing 0/0 host succeeded which was confsing.