13105 - MC::UnixDaemon should remove the pid file it created on exit
The daemon when configured to write a pid would not remove it when it
was done.

Rather than try to catch all exceptions specifically and install trap
handlers with exit logic sprinkled all over we now have a single rescue
block in the daemon that handles that and the UnixDaemon class will
remove the pid file when its done.