13615 - The SSL security plugin should use the MCollective::SSL class
The SSL plugin predates the MCollective::SSL class but the intent is
that all crypto in the framework be done in the M::SSL class so that
testing and bug fixes in this critical code is contained.

This commit signature verification logic to the M::SSL class with tests
and adjust the SSL security plugin to use the M::SSL class for its
crypto needs.  The SSL security plugin gains access to the fixes for the
Ruby core SSL key reading bugs via the same