14159 - Make Rakefile more robust for rpms
Previously the Rakefile didn't recognize Fedora as an lsbdistro and wouldn't
try to use rpmbuild-md5. This commit adds a check for rpmbuild-md5, which is
then preferred over rpmbuild because it is backwards compatible with older ELs.
It also adds recognition for Fedora as an lsbdistro so if src rpms are build on
Fedora, the rake rpm task will succeed.