Init script cleanup, dead code revival
Previously, the echoing after start, stop and restart would never be reached
because of the exits that were called before them. Similarly, the `touch $lock`
was never reached, which is good because $lock wasn't previously defined. This
refacters the dead code to be run before either of the exits are called. This
also modifies force-reload to run a restart on failure.