14960 - MCollective client discovery randomize should take a optional seed.
The logic to select a random selection of nodes was wrong, it would
delete from the Client#discover result which was @discovered_agents
thus actually modifying the clients concept of how many nodes were
discovered.

This commit adjusts the code style a bit and ensure we modify a clone of
the discovered hosts and not the actual list.

Also add a specific option --limit to the --limit-nodes flag to retain a
common used shortcut which now failed with the addition of --limit-seed

Add test coverage for the pick_nodes_from_discovered method and make it
public rather than private