15031 - Add summary methods to the DDL and RPC Results
Agent DDL actions can now contain a summarize block which will apply client
side functions to the response data.

These functions are implemented as aggregate plugins and it is only
necessary for them to be present client side.

Example usage :
  summarize do
    aggregate average(:foo)
  end

In the example 'average' is a client side function that will determine the
average of all of the :foo values in the response hash. Parameters
passed to the client side functions have to be outputs defined in the
ddl actions.