Made request id an attribute reader for Client
This is to enable the reading of the request id in a standard RPC
request so that replies can be matched with their requests if they are
performed in a "detached" manner.

When ::Client is initialized in ::RPC::Client, it is made an
attr_reader, so in an instance of ::RPC::Client

mc = rpcclient(rpc)
mc.client.reqid                     #=> nil

mc.action()
mc.client.reqid                     #=> returns request id.