15474 - when loading agents try to instantiate an instance of each and validate it works
Agents that fail to load like ones that gets a DDL failure from the RPC
system still showed up in the plugin list and could theoretically be
instantiated by things like rpcutil agent agent_inventory.

We now try to instantiate an instance of each agent at load time and if
we cant we kick it out of the agent list and log errors.  This is not a
perfect test but will catch any issues in the constructor which seems to
be the norm so far for big scale validations