15109 - Refactor DDL to be a class per type of DDL
The DDL for all different plugins used to be contained in one big class,
this splits that class out into a base and several sub classes that
captures the behavior for individual plugin types.

The base standardizes the metadata, input and output so that these stays
constant throughout the various types of DDL.

The old DDL class has been turned into a module instead of a class but
unfortunately no good way could be found to not require all existing DDL
related code to change other than providing a custom #new method on this
module that is a factory for the specific DDL classes.  While not ideal
the only improvement we'd have made would be to call this factory
something else, the basic behavior would have stayed the same.  With
this factory in place all existing DDL related code in applications,
sample web apps etc seem to Just Work.

Tests have also been similarly split.