16569 - Timeout from DDL does not propagate when using direct discovery
Each Message object has options, filters, timeouts and such.  The client
should use the options in the Message object when ever it gets asked to
publish a request - else take what comes in as an arg or already there.

As the RPC framework is now completely based on the Message object we
did not set it in the constructor of the RPC client rather relying on
the per message behavior as that will be less prone to failure

Ideally we should completely revisit how the client class handles
options but this does improves things in general - the stats method
will be correct in all cases for the most recent message where
previously there was a small race

maint: fix a test that fails with --order random