17138 - rpcutil#get_data always pass input to the data plugins even when none is expected
Previously all data plugins required an input even those that would not
logically do anything with that input.

Data plugins should though be able to have no input and this bug was
fixed in 2.2.1 but unfortunately the rpcutil#get_data action would in
the case of no input turn the nil into "" which means get_data could not
be used to retrieve data from such plugins.

This fixes that short coming, also removes some unneeded input
validation that is now handled by the DDL before invoking the agent