16414 - Log action failures
When an action fails for unknown reasons we propagate the error back to
the client but not the full backtrace.  We also never logged the full
backtrace which made it very hard to debug failures in actions.

We now log on the server backtraces for unknown failures.

Also provide test coverage for the RPC::Agent#handlemsg method