16018 - mcollectived should expand members of the libdir to full paths
We accepted relative libdirs in the server and client config files
which would lead to unexpected results depending on the working
directory when launching the program.  Worse when daemonizing we always
change the working directory which means the absolute paths will never
work.

The only sane thing we can really do is only allow absoluate paths and
this change enforces this now.