19734 - Make default port 61613 in connector plugins, to match convention/docs
Docs and comments for the ActiveMQ connector stated that the port would default
to 61613 (the closest thing to a standard Stomp port we've been able to find) if
unspecified; instead, it defaults to 6163. This commit changes the default port
to match what we said, and does the same in the RabbitMQ connector.