Fix directed request on subcollectives with rabbit connector.
The identity being used in the subscriptions hash for direct connections was
just using the host identity, not the collective also. Thus even though
the core mcollective was asking for subscription to a direct queue in each
sub-collective, they were all being hashed to the same identity, and thus
only the first one (in the list of subcollectives) was being listened to.

This causes the following issues (when using directed addressing):
. Commands using subcollectives (e.g. mco puppet status -T ldn) were failing
  to get replies (as the mco agents never saw the requests in the subcollective
  directed queue).
. Agents with a sub-collective list where the main collective was not the first
  entry never saw direct requests, as they only ever subscribed to the first
  queue for that subcollective