Docs: Adjust a bunch of per-character intra-emphasis escaping
Individual _, *, and [ characters in words were sometimes being escaped with a
one-character code span. It's generally better to just wrap the whole word in a
code span, so this commit adjusts that as needed. This is part of a minor
cleanup across the docs to prepare for switching Markdown libraries.