22220 - add an install.rb file to mcollective
facter, hiera, and puppet all have an install.rb file, which does basic
installation of the project, with ruby library prereq checking, default config
and library file placement, and she-bang fixes for the ruby.  This simplifies
packaging, because the install.rb can be called in the install phase of the
various packaging methods for file placement instead of duplicating the logic
of copying files around.  This commit adds an install.rb to mcollective, ported
from the facter install.rb. It installs the .cfg.dist files as .cfg files in
$confdir, and plugins in /usr/libexec/mcollective/mcollective by default. This
install.rb does not support windows, because install.rb patterns are quite
different on Windows (and so is MCollective, I believe), and also because
Windows is foreign to me and I'd probably massacre it.