22361 - Mcollective plugins cannot express dependencies
In the past it was impossible to express system specific dependencies
while building packages with the plugin packager.

This commit adds the ability to specify system specific dependencies by
adding either the 'debian::' or 'redhat::' prefix to a dependency.

Examples:

  CLI
    mco plugin package --dependency="debian::ruby-net-ping"
                       --dependency="redhat::rubygem-net-ping"

  buildops.yaml
    dependencies:
      - "debian::ruby-net-ping"
      - "redhat::rubygem-net-ping"

This will add the ruby-net-ping dependency when the deb packager is
invoked and the rubygem-net-ping dependency when the rpm packager is
invoked.