(#17067) Add packaging data files to ext
This commit adds ext/build_defaults.yaml and ext/project_data.yaml, both of
which define the targets, files to be installed, and package types to be
installed. This allows our standard packaging tools to work.

This also updates the .gitignore to include the ext/packaging directory, where
the packaging tools will be cloned.