(#17067) Simplify mcollective.install file
Previously the mcollective.install file specified the destination directory,
even though it didn't differ from the source. It also installed the init script
directly. This commit removes the destination half of the file, as it is
implied in the source pathing. It also removes the init script installation,
which happens automatically because of how the file is named in the debian
directory.