(#17067, #14612, #19319) Templatize and improve the mcollective.spec
Our packaging tools expect to find a redhat spec template in ext/redhat. This
commit copies the current mcollective specfile into an erb file. It also brings
in the work of https://github.com/puppetlabs/marionette-collective/pull/41 to
add systemd files for the systemd init system and updates the specfile to use
them on systems with systemd.  Thanks to Steve Traylen (traylenator on github)
for the systemd work included here.