(#5924, #14206) Update debian packaging to use quilt instead of dpatch
dpatch is a long deprecated patch mechanism for debian. It is now a lintian
error instead of a warning. This commit updates the build to use quilt instead
of dpatch. It also removes unneeded patches that weren't even being applied and
also add the source/format file to indicate the package uses the quilt
standard. The commit also applies the initlsb.dpatch patch to the debian init
script for debian.