(#17067) Update Rakefile to invoke new tasks
This commit updates the logic of the Rakefile to maintain backward
compatibility with the old build tools. Now the top level tasks like deb, rpm,
gem, and doc simply wrap invocations of package:deb, package:tar, package:rpm,
package:gem, and package:doc. The pkg directory (where packages built using the
new tools go), is then moved to the build directory to maintain compatability.
This allows the same tasks to be invoked with the same results. There are some
caveats.
  - First, only the RELEASE can be specified via environment variable. The version
    can be changed by tagging the project and the name can be changed by editing
    ext/project_data.yaml.
  - Second, the version is no longer static/determined by constants in the
    rakefile. It is linked to the git describe, which is based on the last
    signed tag.