(#17067) Refactor mcollective.spec.erb to use install.rb
Previously the specfile replicated much of the logic in the Makefiles that got
ported into install.rb. This commit moves the specfile to use the install.rb
instead of doing the same work manually. Using install.rb means that certain
prereqs are checked, so BuildRequires are added for rubygems and stomp. This
also adds the docs to the common package and ensures that mcollectived has sane
permissions.