(maint) Remove build dependencies on rubygems, stomp library
There is no need to have rubygems or the stomp library at build time, but
install.rb was checking that they were there and could be loaded. This commit
removes the check for stomp and rubygems from install.rb and also removes the
corresponding build dependencies from the redhat and debian packages.