(MCO-97) Config class does not parse fixnum config parameters correctly
In the past it was possible to set non integer like values where
integers where expected in config files. Example if publish_timeout =
five then publish_timeout would be set to 0.

This commit changes the way integer like strings are parsed so as to
fail if a non integer value is found.