(MCO-146) mco ping logs bogus warning
When we added to ability to the client to log a warning when responses
received are less than responses expected we ignored the fact that
applications like ping expect 0 responses, meaning that ping would
always log a warning even when it is behaving normally.

The client receiver will now check if expected responses is greater than
actual responses before logging.