(maint) Remove version_strategy override
Previously, the version_strategy needed to be set at odd_even so that the
versioning strategy of mcollective would be reflected when building and
shipping a release. Now that mcollective is moving to semver with release
candidates, this setting can be removed, as semver with release candidates is
the default value.