(MCO-173) Improve logging when using STOMP 1.1 features with a gem that only supports STOMP 1.0
In the past when a stomp protocol incompatibility error occurred during
runner startup the daemon would not start and no error message would be
logged. This would however always happen if a failure occurred during
initialization of the runner.

This commit adds an exception handler to runner#initialize that will
make sure that an error message is logged when it fails to start.