MCO-198 Connectors should not suggest heart-beats if the installed gem cannot do them.
This commit adds a new helper to the rabbitmq and activemq connectors
`stomp_version_supports_heartbeat?` which allows us to branch more
of the behavior around if the gem can do heartbeating.  We then use
this to not prompt you to enable STOMP 1.1 heart-beats if the
installed gem cannot cope with it.

It also softens the suggestion from a WARN to an INFO, as the user might not
want to enable them, especially if they are using activemq < 5.9 (or whatever
version of rabbitmq adds mature support).