MCO-191 Log receipt and contents of non-MESSAGE frames
Under normal operations @connection.receive is expected to return a
Stomp::Message object containing a MESSAGE frame.  If we get some other frame,
or no frame at all we still pass that up the layers and decode it, which can
lead to some quite dense error messages from the security provider layer.

Here we change the behaviour of the activemq and rabbitmq connectors
a little, to log these frames of unexpected type at WARN level,
with an inspection of the message as a DEBUG message.