Split agent reply timeout from execution timeout
If the agent's action times out in the middle of sending a reply, it may
send a partial message over the wire, which can lead to errors in the
messaging broker.

Run the agent's action in its own timeout block, and if it finishes,
yield the replies in a separate timeout block. This way, it should be
less likely that a partial message is sent out over the wire.