(MCOP-17) Add dist macro to Release field
When building the same plugin for multiple platforms at once, it is important
for the NVR of those packages to be unique for certain deployment platforms
like satellite. Previously the release did not include the dist macro, which
is the easiest way to ensure a unique NVR across platforms. This commit adds
the dist macro to the Release field. This means that instead of having an NVR
like mco-plugin-1.2.1-1 everywhere, now the NVR will be
mco-plugin-1.2.1-1.fc19, mco-plugin-1.2.1-1.el5, etc. The leading ? in the
macro means that if undefined, nothing will be printed, but it won't be an
error.