(MCO-188) Update rubygems dependency to 1.3.7 or greater
Mcollective will fail to start if certain very old versions of rubygems are
employed. Those versions happen to be readily available on el5. Fortunately,
yum.puppetlabs.com offers version 1.3.7 of rubygems, which alleviates this
problem. This commit updates the requires for rubygems to specify a minimum
version of 1.3.7, to ensure that once installed mcollective will actually be
able to start.