(MCO-262) Improve line parsing in flatfile discovery
In the past flatfile discovery would fail if a line started started or
ended with whitespaces or if a empty line was encountered.

This commit softens the definition of what a valid line is by now
stripping leading and trailing whitespace characters before validating
the line with the regular expression.

Empty lines and lines starting with the comment character # will now
also be skipped.