MCO-285 Fix for `mco plugin package` on RedHat variants
The plugin application failed to include the RedHat label to
the package name.

This commit fixes the above issue by correctly adding the
version label to the package name; we now assume that the
label always starts with a dot, as documented in:
http://fedoraproject.org/wiki/Packaging:DistTag#Possible_values_for_.25.7Bdist.7D

Also, to make more robust the way we retrieve the list of
packages paths, we now use glob for both the rpmdir and srpmdir
directories and replace the label part of the file name with a
wildcard.

The fix was based on the patch suggested by Daniel Schleenzig.