(MCO-386) MCO README.md should express semver
The puppet readme (https://github.com/puppetlabs/puppet/blob/master/README.md)
explains the maintenance we have for versions of puppet and that it is
semantically versioned. We recently encountered an issue with Facter where
users were surprised that we did not ship a security release for a previous
major version of facter. The previous assumption may have been that the puppet
README was sufficient for all of our FOSS projects, but this is probably not
the case. We should update the README.md in mcollective with the same verbage as
puppet so our intended release lifecycle is clear.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>