(MCO-407) Add requirement for connector plugins to have ddl files.
Due to the primary connector plugins being mainly shipped with core and
developed by Puppet Labs fixing bugs or extending the connectors often
require changes in core to happen at the same time. This puts us in the
position where connector plugins often depend on specific minimum
version of MCollective but that we cannot enforce.

This commit adds the requirement that connector plugins should always
ship with a ddl.

(MCO-406) Add DDL's to Connectors

This commit adds ddl's to both the ActiveMQ and RabbitMQ connectors and
adds test coverage to the MCollective::Connector::Base class.