MCO-363 Add structured matching semantics to has_fact?
Here we add simple structured fact support to the has_fact? utility
method.

In order to get there we extract the core matching into test_fact_value,
to allow for it to be applied to structured data compound data
structures more readily.  Then we add the semantics for matching
against structured values (any of the the contents of an Array or
keys of a Hash should match the test)