(MCO-205) Fix call_agent_batched to work with active_record
MCollective monkey patches array to include #in_groups_of, unless the
method is already defined. In the case of using MCollective and
active_record this method is already defined.

MCollective's version of #in_groups_of can yield multiple results to the
block parameter. This is not the same behaviour as ActiveRecord's
implementation so when the method gets called using MCollective's
symantics #in_groups_of does not show the expected behaviour.

This commit is based on a recommendation by Fabian Ruff and changes the
invocation of #in_groups_of to match that of the method defined by
ActiveRecord.