MCO-411 remove examples of catching RPCErrors in response blocks
It was originally intended that the appropriate exception for a
given rpc response would be propogated into the block supplied to
the call_agent method.  This isn't possible in ruby as noted in the
discussion around MCO-264.

Here we rework the misleading example code to reflect that we no
longer attempt to re-raise the exception in the supplied block, and
instead demonstrate testing response[:body][:statuscode].