MCO-416 specify the --daemonize option in init scripts
MCO-181 added the --daemonize and --no-daemonize options to mcollectived so
that the invoking init script could have final say about what it expects the
mcollectived to do - to daemonize or stay in the foreground - and so insulate
itself from a potential mismatch in the server.cfg and the expected behaviour.

Here we update the init scripts we have to make use of this feature.