(MCO-68) Allow --batch flag to use percentages
This commit adds the ability to RPC::Client::call_agent_batched to
determine a batch size based on a percentage of the amount of hosts
discovered.

This is achieved by softening the argument validation of the --batch
flag to also allow strings which match the regular expression /^\d+%?$/.

The ability to override and reset global batch sizes has been kept.