(MCO-408) Add ability to globally define if agents should be loaded
This commit adds the global configuration option, 'activate_agents',
which functions similarly to 'activate_agent' which can be specified
on a per agent bases.

When set to false no agents will be loaded. This behaviour can be
overridden on a per agent basis.

The current default behaviour of loading all agents by default has
been kept.