MCO-486 expose SSL cipher to connector parameters
Here we add options to activemq and rabbitmq connectors to specify the desired
set of OpenSSL ciphers when communicating with the middleware.

As an example, this would choose ciphers from the TLSv1.2 suite, while
excluding some known weak ciphers.

     plugin.activemq.pool.1.ssl.ciphers = TLSv1.2:!MD5:!LOW:!EXPORT

See the OpenSSL documentation for an explanation of these cipher strings.
https://www.openssl.org/docs/apps/ciphers.html#CIPHER_STRINGS