(MCO-466) Add search of $libdir/mcollective/agent/$agent/$action to implemented by
The commit d9d4354 implemented a default search path for scripts
supplied as parameters to #implemented_by. However the path did not
correspond with the path described in the feature ticket
http://projects.puppetlabs.com/issues/13064

This commit adds the previously described search path with the aim of
deprecating the incorrect search path in the future.