MCOP-519 pin to version 0.28.x of rubocop
The rubocop project releases farily frequently adding new style
rules, so pin to an agreed upon version to avoid sudden rework when
a new version of rubocop ships.

We need to adopt a recurring task to upgrade and evaluate new rubocop
rules, this commit should not be taken as following the dangerous
precedent of pinning and never updating a depedency.