MCO-580 replace uses of autoload with require
The codebase was making heavy use of autoload in order to defer compilation of
namespaces until it was known for certain that they were used.  This behaves
poorly in a multithreaded environment as noted here
https://www.ruby-forum.com/topic/3036681#1032667 and observed in
https://tickets.puppetlabs.com/browse/MCO-580

Here we switch the autoloads for requires.  This incurs a small cost in terms
of initial compile time and memory footprint, but the tradeoff seems fair in
order to avoid the rb_bug we can see due to the interaction of autoload and
threads.

Before (master @bc474f3):

    $ ruby -Ilib -rmcollective -e'puts `ps -o rss= -p #{Process.pid}`'
    12960

    $ time ruby -Ilib -rmcollective -e1

    real    0m0.147s
    user    0m0.122s
    sys     0m0.021s

With this patch:

    $ ruby -Ilib -rmcollective -e'puts `ps -o rss= -p #{Process.pid}`'
    15876

    $ time ruby -Ilib -rmcollective -e1

    real    0m0.199s
    user    0m0.169s
    sys     0m0.027s

So that's about 2.6MB added to the initial startup size and 0.05 seconds to the
startup time of mco or mcollectived.

Measurements taken on a Mid 2012 MacBook Air. ruby 1.9.3p551