MCO-315 make $libdir prepend values to $LOAD_PATH
Currently the semantics for $libdir are that it's an array of paths
that MCollective should search for plugins.  The change we make
here is to make $libdir more generally extend $LOAD_PATH and use
$LOAD_PATH for all plugin loading.

This allows plugins for to be located in gems and for plugin
resolution to be influenced by the RUBYLIB environent variable, or
the ruby -I switch.

In the configuration parser we make the values supplied by the user
be prepended to the ruby $LOAD_PATH.

When loading plugins we now search all of $LOAD_PATH (via
Config.instance.libdir) for ddl searches and plugin loading.