MCO-587 fix client gem loading
In MCO-580 6bdc58f97 we changed all the autoloads for requires, overlooking
that the mcollective-client gem witholds `lib/mcollective/runner.rb`.  This
causes the gem to to unusable as the require of 'mcollective/runner' can not be
resolved in a clean install.

Here we move the require out to the only users of MCollective::Runner, its unit
tests and the mcollectived script.