(maint) specify a default connection_timeout
If the middleware is down, or unreachable (which it's very likely to be for
stomp1:6163 as it's just placeholder) the client will just sit and wait which
is a poor first impression.  Add connection_timeout to the example config file,
so a user won't sit wondering why `mco ping` in an unconfigured environment
simply waits.