(MAINT) Add mco_master role to acceptance
This commit adds an `mco_master` role to the beaker host files,
as well as using this role in the tests and pre-suit files to
determine where to deploy the mco-broker infrastructure.

Since the `mco_master` role can be applied to any node in the
hosts lists, the `server.cfg` file has been added to the
`060_Install-activemq.rb` pre-suite step as a parameterized
HEREDOC string.