(MCOP-521) mco runonce with powershell provider test
This commit adds an acceptance test to verify that executing
a `mco puppet runonce` where a Windows agent is classified
with an exec resource that uses the powershell provider correctly
executes on the Windows system.

This test is currently expected to fail until MCOP-521 is resolved.
The test assertion is wrapped in an `expect_failure` method to
prevent the test run from turning red until MCOP-521 has been fixed.
At that time, the test will fail and the wrapper can be removed to
flip the result.